Move disease model classes into `hpotk.annotations` module.
