+++ src/hpotk/__init__.py
@@ -5,4 +5,5 @@ from . import constants
 from . import graph
 from . import ontology
 from . import algorithm
+from . import annotations
 from . import util
