+++ src/hpotk/annotations/__init__.py
@@ -1,0 +1,5 @@
+from ._base import Ratio, EvidenceCode, Sex, AnnotationReference
+from ._base import HpoDiseaseAnnotation, HpoDisease
+from ._base import SimpleRatio
+
+from . import load
