+++ src/hpotk/annotations/_base.py
