+++ src/hpotk/model/__init__.py
@@ -3,7 +3,6 @@ import typing
 from ._term_id import TermId
 from ._base import Identified, Named
 from ._term import MinimalTerm, Term
-from . import disease
 
 # Types
 ID = typing.TypeVar('ID', bound=TermId)
