--- src/hpotk/model/disease/__init__.py
@@ -1,5 +1,0 @@
-from ._disease import HpoDiseaseAnnotation, HpoDisease
-from ._disease import EvidenceCode, Sex, Ratio, AnnotationReference
-from ._disease import SimpleRatio
-
-from ._simple import SimpleHpoDisease, SimpleHpoDiseaseAnnotation
